backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
alembic==1.14.0
asyncmy==0.2.10
redis==5.2.1
orjson==3.10.13
python-dotenv==1.0.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
Pillow==10.4.0
